Added support for retrieving different kinds of ranking.
Search Results are now improved, containing in-depth results such as favorites, full synopsis, short synopsis, author, genres and etc. (this applies to the rankings, as well).

You can now search for individual rankings by using the getRanking() method on Amatsuki class, it will ask for Rankings enum and reply back with the top 15 of the ranking specified.

FrontPagePanels are now deprecated, or rather removed as per the change of Trending's base code which now implements getRanking().

You can now select the Time Order of the ranking as well, either Weekly or Daily, etc... by using the OrderBy enum.